+++ shuup/admin/forms/_auth.py
@@ -7,8 +7,12 @@
 # LICENSE file in the root directory of this source tree.
 from django.contrib.auth import get_user_model
 from django.contrib.auth.forms import AuthenticationForm
+from django.contrib.auth.tokens import default_token_generator
 from django.core.exceptions import MultipleObjectsReturned, ObjectDoesNotExist
+from django.db.models import Q
 from django.utils.translation import ugettext_lazy as _
+
+from shuup.core.utils.forms import RecoverPasswordForm
 
 
 class EmailAuthenticationForm(AuthenticationForm):
@@ -43,3 +47,29 @@ class EmailAuthenticationForm(AuthenticationForm):
             return getattr(user_by_email, user_model.USERNAME_FIELD)
 
         return username
+
+
+class RequestPasswordForm(RecoverPasswordForm):
+    token_generator = default_token_generator
+    subject_template_name = "shuup/admin/auth/recover_password_mail_subject.jinja"
+    email_template_name = "shuup/admin/auth/recover_password_mail_content.jinja"
+    from_email = None
+
+    def __init__(self, *args, **kwargs):
+        self.request = kwargs.pop("request")
+        super(RequestPasswordForm, self).__init__(*args, **kwargs)
+
+    def save(self):
+        user_model = get_user_model()
+        username = self.cleaned_data.get("username")
+        email = self.cleaned_data.get("email")
+        username_filter = {
+            "{0}__iexact".format(user_model.USERNAME_FIELD): username
+        }
+        # only staff and active users
+        active_users = user_model.objects.filter(
+            Q(is_active=True, is_staff=True),
+            Q(**username_filter) | Q(email__iexact=email)
+        )
+        for user in active_users:
+            self.process_user(user)